standardized response across controllers

+++ Controllers/DrugCtrls.js
@@ -13,12 +13,21 @@ export const handleCreateDrug = async (req, res, next) => {
     const drugExists = await Drugs.findOne({ genericName, brandName });
 
     if (drugExists) {
-        return res.status(400).json({ message: "Drug with the same generic name and brand name already exists" });
-    }   
+    const error = new Error(
+        "Drug with the same generic name and brand name already exists"
+    );
+    error.statusCode = 400;
+    return next(error);
+}  
 
     const newDrug = await Drugs.create({genericName,therapeuticClass,indications,contraindications,sideEffects,interactions,route,brandName,manufacturer,dosageForm,strength,nafdacNumber,costPrice,sellingPrice,minimumStockLevel,isActive});
 
-    res.status(201).json(newDrug);
+    return successResponse(
+    res,
+    201,
+    "Drug created successfully",
+    newDrug
+);
   } catch(error){
     next(error)
   }
@@ -27,10 +36,12 @@ export const handleCreateDrug = async (req, res, next) => {
 
 export const handleGetAllDrugs = async (req, res, next) => {
     try{const drugs = await Drugs.find();
-    res.status(200).json({
-    message: "List of drugs retrieved successfully",
+    return successResponse(
+    res,
+    200,
+    "List of drugs retrieved successfully",
     drugs
-});
+);
 } catch(error){
     next(error)
     }
@@ -51,10 +62,12 @@ export const handleGetDrugById = async (req, res, next) => {
         error.statusCode = 404;
         return next(error);
     } 
-    res.status(200).json({
-    message: "Drug retrieved successfully",
+    return successResponse(
+    res,
+    200,
+    "Drug retrieved successfully",
     drug
-});; 
+); 
     } catch(error){
         next(error)
         
@@ -77,11 +90,15 @@ export const handleGetDrugsByBrandAndGenericName = async (req, res, next) => {
         query.genericName = { $regex: new RegExp(genericName, "i") };
     }
     const drugs = await Drugs.find(query);
-    res.status(200).json({
-    message: "Drugs retrieved successfully",
-    count: drugs.length,
-    drugs
-});
+    return successResponse(
+    res,
+    200,
+    "Drugs retrieved successfully",
+    {
+        count: drugs.length,
+        drugs
+    }
+);
     } catch(error){
         next(error)
     }
@@ -93,27 +110,27 @@ export const handleUpdateDrug = async (req, res, next) => {
         const { id } = req.params;
     const { genericName, therapeuticClass, indications, contraindications, sideEffects, interactions, route, brandName, manufacturer, dosageForm, strength, nafdacNumber, costPrice, minimumStockLevel, isActive } = req.body;  
     
-    const drug = await Drugs.findById(id);
-
-    if (!mongoose.Types.ObjectId.isValid(id)) {
-    const error = new Error("Invalid drug ID");
-    error.statusCode = 400;
-    return next(error);
-}
-
+    
 if (!mongoose.Types.ObjectId.isValid(id)) {
     const error = new Error("Invalid drug ID");
     error.statusCode = 400;
     return next(error);
 }
 
+const drug = await Drugs.findById(id);
+
         if (!drug) {
             const error = new Error("Drug not found");
             error.statusCode = 404;
             return next(error);
     }
-    const deletedDrug = await Drugs.findByIdAndUpdate(id,  {genericName, therapeuticClass, indications, contraindications, sideEffects, interactions, route, brandName, manufacturer, dosageForm, strength, nafdacNumber, costPrice, minimumStockLevel, isActive,sellingPrice}, { new: true });
-    res.status(200).json(updatedDrug);
+    const updatedDrug = await Drugs.findByIdAndUpdate(id,  {genericName, therapeuticClass, indications, contraindications, sideEffects, interactions, route, brandName, manufacturer, dosageForm, strength, nafdacNumber, costPrice, minimumStockLevel, isActive,sellingPrice}, { new: true });
+    return successResponse(
+        res,
+        200,
+        "Drug updated successfully",
+        updatedDrug
+    );
     }catch(error){
         next(error)
     }
@@ -123,17 +140,28 @@ if (!mongoose.Types.ObjectId.isValid(id)) {
 export const handleDeleteDrug = async (req, res, next) => {
     try{
         const { id } = req.params;
+
+if(mongoose.Types.ObjectId.isValid(id)){
+    const error = new Error("Invalid drug ID");
+    error.statusCode = 400;
+    return next(error);
+}
+
     const drug = await Drugs.findById(id);
+
     if (!drug) {
         const error = new Error("Drug not found");
         error.statusCode = 404;
         return next(error);
     }
-    await Drugs.findByIdAndDelete(id);
-    res.status(200).json({
-    message: "Drug updated successfully",
-    drug: updatedDrug
-});
+
+   const deletedDrug = await Drugs.findByIdAndDelete(id);
+    return successResponse(
+        res,
+        200,
+        "Drug deleted successfully",
+        deletedDrug
+    );
     } catch(error){
         next(error)
     }
